requirements.txt
kafka-python==2.0.2
apache-flink==1.20.0